+++ src/jast/antlr/JavaLexer.java
@@ -1,4 +1,4 @@
-package antlr;// Generated from JavaLexer.g4 by ANTLR 4.7.1
+package jast.antlr;// Generated from JavaLexer.g4 by ANTLR 4.7.1
 import org.antlr.v4.runtime.Lexer;
 import org.antlr.v4.runtime.CharStream;
 import org.antlr.v4.runtime.*;
