+++ src/jast/antlr/JavaParser.java
@@ -1,4 +1,4 @@
-package antlr;// Generated from JavaParser.g4 by ANTLR 4.7.1
+package jast.antlr;// Generated from JavaParser.g4 by ANTLR 4.7.1
 import org.antlr.v4.runtime.atn.*;
 import org.antlr.v4.runtime.dfa.DFA;
 import org.antlr.v4.runtime.*;
